updated row styling on employee table and fixed bug in comment modal popup

+++ application/views/home_view.php
@@ -10,9 +10,20 @@
 			<th>CLOCK OUT</th>
 			<th>DAILY TOTAL</th>
 		</tr>
+
+		<?php $c = true; ?>
+		
 		<?php foreach($week['entries'] as $entry) { ?>
-         
-		<tr>
+		<?php 
+		if($c) { 
+			$c = false;
+     		$class = 'row_1';
+		} else {
+     		$c = true;
+    	 	$class = 'row_2';
+		} ?>
+
+		<tr class="<?= $class ?>">
 
 			<td><?=date("M d, Y, g:i a", strtotime($entry['start']))?></td>
 			<td>
@@ -33,9 +44,9 @@
 			</td>
 		</tr>
 
-		<tr>
-			<tr>	
-				<td colspan="3">
+		<tr class="<?= $class ?>">
+
+			<td colspan="3">
 				<div class="comment">
 					<?php
 					if($entry['end'] == NULL) { 
@@ -43,29 +54,28 @@
 					} else {
 						if($entry['comment']) { ?> 
 						<?= $entry['comment'];?><br><br>
-						<a href="#insert_comment<?=$entry['id']?>" class="modal_popup">
-						<i class="icon-pencil"></i> Edit Comment</a> | 
-						<a href="/index.php/time_entry/delete_comment/<?=$entry['id']?>" 
-						onclick="return confirm('Are you sure you want to delete this employee?')">
-						<i class="icon-remove"></i> Delete Comment</a> 
+						<a href="#insert_comment<?=$entry['id']?>" class="modal_popup"><i class="icon-pencil"></i> Edit Comment</a> | 
+						<a href="/index.php/time_entry/delete_comment/<?=$entry['id']?>" onclick="return confirm('Are you sure you want to delete this comment?')"><i class="icon-remove"></i> Delete Comment</a> 
 						<?php } else { ?>
 						<a  href="#insert_comment<?=$entry['id']?>" class="modal_popup"><i class="icon-star"></i> Add a Comment</a>
 						<?php } 
 					} ?>
-				</td>
-			</div>
-			</tr>
-			</div>
+					<div class = "hidden">
+				   	<?php $placeholder = "placeholder='Enter a comment about what you did today'"; ?>
+				   	<?= form_open('time_entry/insert_comment', $attributes = array('id' => 'insert_comment' . $entry['id']) ); ?>
+				   		<?= form_hidden('id', $entry['id']) ?>
+				   		<label for "Daily Comment" </label>
+				   		<?= form_textarea('comment', '', $placeholder); ?>
+				   		<?= form_submit('submit', 'Submit'); ?>
+				   	<?= form_close(); ?>
+   					</div>
+				</div>
+
+			</td>
+
 		</tr>
-		<div class="hidden">
-		   	<?php $placeholder = "placeholder='Enter a comment about what you did today'"; ?>
-		   	<?= form_open('time_entry/insert_comment', $attributes = array('id' => 'insert_comment' . $entry['id'], 'input type' => 'text') ); ?>
-		   		<?= form_hidden('id', $entry['id']) ?>
-		   		<label for "Daily Comment" </label>
-		   		<?= form_textarea('comment', $entry['comment'], $placeholder); ?>
-		   		<?= form_submit('submit', 'Submit'); ?>
-		   	<?= form_close(); ?>
-		</div>
+		
+	
 	<?php } ?>
 
 </table>
